Move test report module
