+++ src/pages/test-report/components/eta-invalid-modal/eta-invalid-modal.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { EtaInvalidModal } from './eta-invalid-modal';
-import { TestReportModalModule } from '../test-report-modal.module';
+import { TestReportModalModule } from '../test-report-modal/test-report-modal.module';
 
 @NgModule({
   declarations: [
