+++ src/pages/test-report/components/legal-requirements-modal/legal-requirements-modal.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { LegalRequirementsModal } from './legal-requirements-modal';
-import { TestReportModalModule } from '../test-report-modal.module';
+import { TestReportModalModule } from '../test-report-modal/test-report-modal.module';
 
 @NgModule({
   declarations: [
