+++ src/pages/test-report/components/test-report-components.module.ts
@@ -22,7 +22,7 @@ import { TimerComponent } from './timer/timer';
 import { EndTestModalModule } from './end-test-modal/end-test-modal.module';
 import { LegalRequirementsModalModule } from './legal-requirements-modal/legal-requirements-modal.module';
 import { EtaInvalidModalModule } from './eta-invalid-modal/eta-invalid-modal.module';
-import { TestReportModalModule } from './test-report-modal.module';
+import { TestReportModalModule } from './test-report-modal/test-report-modal.module';
 
 @NgModule({
   declarations: [
